Updates per review comments
