+++ frontend/packages/topology/src/components/CreateConnectorWidget.scss
