+++ frontend/packages/topology/src/components/DefaultCreateConnector.scss
@@ -1,17 +1,25 @@
 .topology-default-create-connector {
+  --create-connector-color: var(--pf-global--active-color--400);
+  --create-connector-bg-color: var(--pf-global--BackgroundColor--light-100);
+
   .topology-connector-arrow {
-    stroke: var(--pf-global--active-color--400);
-    fill: var(--pf-global--active-color--400);
+    stroke: var(--create-connector-color);
+    fill: var(--create-connector-color);
   }
   &__line {
-    stroke: var(--pf-global--active-color--400);
+    stroke: var(--create-connector-color);
     stroke-width: 2px;
     stroke-dasharray: 5px, 5px;
   }
 
-  &__create > * {
-    stroke: var(--pf-global--active-color--400);
-    fill: #fff;
-    stroke-width: 2px;
+  &__create {
+    &__bg {
+      stroke: var(--create-connector-color);
+      fill: var(--create-connector-bg-color);
+      stroke-width: 2px;
+    }
+    &__cursor {
+      fill: var(--create-connector-color);
+    }
   }
 }
